rename database to dal
